refactor: tratar erros turma

+++ backend/src/main/java/com/denisson/backend/domain/useCases/turmas/DeleteTurmaUseCase.java
@@ -1,5 +1,7 @@
 package com.denisson.backend.domain.useCases.turmas;
 
+import com.denisson.backend.domain.exceptions.NotFoundException;
+import com.denisson.backend.domain.models.Turma;
 import com.denisson.backend.domain.repositories.ITurmaRepository;
 
 public class DeleteTurmaUseCase {
@@ -11,6 +13,10 @@ public class DeleteTurmaUseCase {
     }
 
     public void execute(Long id) {
+        Turma existsTurma = repository.findById(id);
+        if (existsTurma == null) {
+            throw new NotFoundException("Turma não encontrada com id " + id);
+        }
         repository.delete(id);
     }
 }
